adição de dados json e arquivo main.js cpy como backup
