+++ codigo/Prototipo/Calendario_sin.html
@@ -135,8 +135,29 @@
     </main>
     
 
+<!--Footer das informações do site
+  <footer>
+    <div class="footer">
+      <div class="row g-0 text-center">
 
-
+        <div class="col-6 col-md-4 p-2">
+          <h5>Seja Nosso cliente</h5>
+          <a href='Cadastro_Med.html' style="color: white;">Cadastre-se</a>
+        </div>
+        <div class="col-6 col-md-4 p-2">
+          <h5>Fale com a gente</h5>
+          <p>Enviar Mensagem</p>
+          <p>Suporte (31) 9 9999-9999</p>
+          <p>Reclame aqui</p>
+        </div>
+        <div class="col-6 col-md-4 p-2">
+          <h5>Criadores</h5>
+          <a href="https://github.com/ICEI-PUC-Minas-PMGES-TI/pmg-es-2024-1-ti1-2401100-g7-saude-e-bem-estar"style="color: white;">Equipe de Diw G7 - Puc Minas Coração Eucaristico </a>
+        </div>
+      </div>
+    </div>
+  </footer>
+ -->
     <script src="./assets/js/Calendario_sin.js"></script>
 </body>
 </html>
