--- pmg-es-2024-1-ti1-2401100-g7-saude-e-bem-estar/README.md
@@ -1,20 +1,0 @@
-# SANTÉ
-Há muitas pessoas que, por várias razões, negligenciam sua saúde e bem-estar. Algumas delas simplesmente não encontram tempo para visitar um médico, enquanto outras podem enfrentar dificuldades para se deslocar até um consultório. Nossa proposta é desenvolver um sistema que possa acompanhar e realizar verificações diárias de informações relevantes, oferecendo orientações sobre o que fazer em determinadas situações e até mesmo auxiliando na identificação precoce de problemas de saúde, na medida do possível. A inovação que buscamos implementar incluirá a capacidade de prever doenças e oferecer um acompanhamento contínuo da saúde.
-
-## Alunos integrantes da equipe
-
-* Matheus Dias
-* Julia Rocha
-* Lucas Silva
-* Eduardo Maia
-* Cleverson Resende
-
-## Professores responsáveis
-
-* Simone Alves Nogueira
-* Diego Augusto de Faria Barros
-* Daniel Victor Souza Saldanha
-
-## Instruções de utilização
-
-Assim que a primeira versão do sistema estiver disponível, deverá complementar com as instruções de utilização. Descreva como instalar eventuais dependências e como executar a aplicação.
