--- pmg-es-2024-1-ti1-2401100-g7-saude-e-bem-estar/codigo/Prototipo/Histórico.html
@@ -1,45 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Histórico de Saúde do Paciente</title>
-  <link rel="stylesheet" href="style.css">
-</head>
-<body>
-  <header> <h1> SANTÉ</h1>
-
-  <h2>Histórico de Saúde do Paciente</h2>
-  </header>
-  <hr>
-  <div id="healthRecords">
-    <h3>Registros de Saúde</h3>
-    <ul id="recordsList"></ul>
-  </div>
-
-  <div id="addRecordForm">
-    <h3>Adicionar Novo Registro</h3>
-    <form id="recordForm">
-      <label for="date">Data:</label>
-      <input type="date" id="date" name="date" required>    
-      <label for="time">Horário:</label>
-      <input type="time" id="time" name="time" required>    
-      <label for="bloodPressure">Pressão Arterial:</label>
-      <input type="text" id="bloodPressure" name="bloodPressure" placeholder="Ex: 120/80" required>
-      <label for="glucose">Glicose (mg/dL):</label>
-      <input type="number" id="glucose" name="glucose" required>
-      <label for="weight">Peso (kg):</label>
-      <input type="number" id="weight" name="weight" step="0.1" required>
-      <button type="submit">Adicionar Registro</button>
-    </form>
-  </div>
-  <hr>
-  <footer>
-
-    <p>© 2024 SANTÉ. Todos os direitos reservados.</p>
-    <p> Criadores - Equipe de Diw G7 - Puc Minas Coração Eucarístico</p>
-    
-  </footer>
-  <script src="script.js"></script>
-</body>
-</html>
