--- pmg-es-2024-1-ti1-2401100-g7-saude-e-bem-estar/codigo/Prototipo/LembreteDeAtividades.html
@@ -1,54 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>SANTÉ</title>
-  <link rel="stylesheet" href="style.css">
-</head>
-<body>
-  <div class="container">
-    <h2>Lembrete Medicamentos</h2>
-    <form id="reminderForm">
-      <label for="activity">Atividades:</label>
-      <input type="text" id="activity" name="activity" required>
-
-      <label for="time">Horário:</label>
-      <input type="time" id="time" name="time" required>
-
-      <button type="submit">Adicionar Lembrete</button>
-    </form>
-
-    <div id="medicineList">
-    <hr> 
-      <h3>Lembrete de Consultas</h3>
-      <form id="medicineForm">
-        <label for="medicine">Medicamentos:</label>
-        <input type="text" id="medicine" name="medicine" required>
-
-        <label for="time">Horário:</label>
-        <input type="time" id="time" name="time" required>
-
-        <label for="date">Data:</label>
-        <input type="date" id="date" name="date" required>
-
-        <button type="button" id="addMedicine">Adicionar Medicamento</button>
-      </form>
-      </ul>
-    </div>
-
-    <div id="reminderList">
-      
-    <hr>
-      <h3>Lembretes</h3>
-      <ul id="activityReminders">
-        <!-- Lista de lembretes de atividades será adicionada aqui -->
-      <ul id="medicineReminders">
-          <!-- Lista de lembretes de medicamentos será adicionada aqui -->
-      </ul>
-    </div>
-  </div>
-
-  <script src="script.js"></script> <!-- Adicionando o script JavaScript -->
-</body>
-</html>
