--- pmg-es-2024-1-ti1-2401100-g7-saude-e-bem-estar/codigo/Prototipo/Login_Med.html
@@ -1,67 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link href="assets/css/main.css" rel="stylesheet">
-    <link href="assets/css/bootstrap.min.css" rel="stylesheet">
-    <title>Santé</title>
-</head>
-<body>
-    <Header>
-        <div class="header">
-          <a href="index.html" class="logo">Santé <img src="assets/img/Saude2.png" alt="saude" class="img-saude"></a>
-          <div class="header-right">
-            <a href="#home">Meu Caso</a>
-            <a href="#contact">Quem Somos</a>
-            <button href="#" type="button" class="btn-verde">Cadastrar</button>
-            <button href="#" type="button" class="btn-verde">Login</button>
-          </div>
-        </div>
-    </Header>
-    
-    <main>
-    <br><div class="container shadow p-3 mb-5 bg-body-tertiary rounded">
-        <form>
-          <h2 class="d-flex justify-content-center">Login de Médico</h2>
-          <div class="form-group">
-            <label for="email">Email:</label>
-            <input type="email" class="form-control" id="email" name="email" placeholder="matheus@gmail.com" required>
-          </div>
-          <div class="form-group">
-            <label for="senha">Senha:</label>
-            <input type="password" class="form-control" id="senha" name="Senha" placeholder="******" required>
-          </div><br>
-          <div class="d-flex justify-content-center">
-            <a href=""><button type="submit" onclick="Login()" class="btn btn-primary">Login</button></a>
-            <a href="index.html" ><button class="btn btn-info">Voltar</button></a>
-          </div>
-          <br><a href="Cadastro_Med.html" class="d-flex justify-content-center link" >Ainda não tem uma conta?</a>
-        </form>
-      </div>
-    </main>
-
-    <footer>
-        <div class="footer ">
-          <div class="row g-0 text-center">
-  
-            <div class="col-6 col-md-4 p-2">
-              <h5>Seja Nosso cliente</h5>
-              <p>Cadastre-se</p>
-            </div>
-            <div class="col-6 col-md-4 p-2">
-              <h5>Fale com a gente</h5>
-              <p>Enviar Mensagem</p>
-              <p>Suporte (31) 9 9999-9999</p>
-              <p>Reclame aqui</p>
-            </div>
-            <div class="col-6 col-md-4 p-2">
-              <h5>Criadores</h5>
-              <p>Equipe de Diw G7 - Puc Minas Coração Eucaristico </p>
-            </div>
-          </div>
-        </div>
-      </footer>
-    <script src="assets/js/main.js"></script>
-</body>
-</html>
