--- pmg-es-2024-1-ti1-2401100-g7-saude-e-bem-estar/codigo/Prototipo/Perfil_Med.html
@@ -1,84 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link href="assets/css/main.css" rel="stylesheet">
-    <link href="assets/css/perfil_med.css" rel="stylesheet">
-    <link href="assets/css/bootstrap.min.css" rel="stylesheet">
-    <title>Santé</title>
-</head>
-<body>
-    <header>
-        <div class="header">
-            <a href="index.html" class="logo"> Santé <img src="assets/img/Saude2.png" alt="saude" class="img-saude"></a>
-            <div class="header-right">
-                <a href="#home">Meu Caso</a>
-                <a href="#contact">Quem Somos</a>
-
-                <div class="dropdown header-right">
-                    <a class="dropdown-toggle" href="#" role="button" data-bs-toggle="dropdown" aria-expanded="false">
-                        <img src="assets/img/perfil.png" alt="saude" class="img-saude">
-                    </a>
-                    <ul class="dropdown-menu">
-                        <li><a class="dropdown-item" href="#">Configuração</a></li>
-                        <li><a class="dropdown-item" href="#">Sair</a></li>
-                    </ul>
-                </div>
-            </div>
-        </div>
-    </header>
-
-    <main><br>
-            <div class="fundo">
-                <div class="">
-                    <img src="assets/img/perfil.png" alt="perfil_grand" align="left" class="img_perfil"><br>
-                    <h1 id="name_perfil">Medico Name</h1><br>
-                    <h4 id="idade_perfil"></h4>
-                    <h4 id="email_perfil"></h4>
-                    <h4 id="telefone_perfil"></h4>
-                <!--<h4 id="cep_perfil"></h4><br> -->
-                    <h4 id="expecializacao_perfil"></h4><br>
-                    <button type="submit" class="btn btn-primary">Editar</button><br>
-                    <button type="submit" class="btn btn-danger">Excluir dados</button><br>
-                </div>
-            </div>
-        <div class=""><br><br>
-            <div class="fundo">
-              <div style="margin-left:20px;">
-                <h1 >Sobre Mim</h1><br>
-                <p>Opcional:</p>
-                <input type="text"><br>
-                <button type="submit" class="btn btn-primary btn-right">Salvar</button>
-                <button type="submit" class="btn btn-primary btn-right">Cancelar</button><br><br>
-              </div>
-            </div>
-        </div>
-    </main>
-
-    <footer>
-        <div class="footer ">
-            <div class="row g-0 text-center">
-                <div class="col-6 col-md-4 p-2">
-                    <h5>Seja Nosso cliente</h5>
-                    <p>Cadastre-se</p>
-                </div>
-                <div class="col-6 col-md-4 p-2">
-                    <h5>Fale com a gente</h5>
-                    <p>Enviar Mensagem</p>
-                    <p>Suporte (31) 9 9999-9999</p>
-                    <p>Reclame aqui</p>
-                </div>
-                <div class="col-6 col-md-4 p-2">
-                    <h5>Criadores</h5>
-                    <p>Equipe de Diw G7 - Puc Minas Coração Eucaristico </p>
-                </div>
-            </div>
-        </div>
-    </footer>
-
-    <script src="assets/js/main.js"></script>
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.6.0/jquery.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/bootstrap/5.3.0/js/bootstrap.bundle.min.js"></script>
-</body>
-</html>
