--- pmg-es-2024-1-ti1-2401100-g7-saude-e-bem-estar/codigo/Prototipo/TelaEscolha.html
@@ -1,74 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link href="assets/css/main.css" rel="stylesheet">
-    <link href="assets/css/bootstrap.min.css" rel="stylesheet">
-    <title>Santé</title>
-</head>
-<body>
-  <!--Cabeçalho do site -->
-    <Header>
-        <div class="header">
-          <a href="index.html" class="logo"> Santé <img src="assets/img/Saude2.png" alt="saude" class="img-saude"></a>
-          <div class="header-right">
-            <a href="#home">Meu Caso</a>
-            <a href="#contact">Quem Somos</a>
-              <button type="button" onclick="window.location.href='Cadastro_Med.html'" class="btn-verde">Cadastrar</button></a>
-              <button href="#" type="button" onclick="window.location.href='Login_Med.html'"class="btn-verde">Login</button></a>
-          </div>
-        </div>
-    </Header>
-
-    
-<!--Conteudo central do site -->
-    <main>
-        <br><br><br><h1 class="text-center">Conecte-se à sua saúde, conecte-se ao futuro, conecte-se conosco.</h1><br>
-
-
-<section id="sectionTelaEscolha">
-    <br>
-    <h1>Queremos te conhecer</h1>
-    <br>
-    <div class="cardTelaEscolha" onclick="window.location.href='pagina1.html'">
-        <h2>Profissional de Saúde</h2>
-      </div>
-      
-      <div class="cardTelaEscolha" onclick="window.location.href='pagina2.html'">
-        <h2>Paciente</h2>
-      </div>
-</section>
-
-
-
-
-<!--Footer das informações do site -->
-<footer>
-    <div class="footer">
-      <div class="row g-0 text-center">
-
-        <div class="col-6 col-md-4 p-2">
-          <h5>Seja Nosso cliente</h5>
-          <p>Cadastre-se</p>
-        </div>
-        <div class="col-6 col-md-4 p-2">
-          <h5>Fale com a gente</h5>
-          <p>Enviar Mensagem</p>
-          <p>Suporte (31) 9 9999-9999</p>
-          <p>Reclame aqui</p>
-        </div>
-        <div class="col-6 col-md-4 p-2">
-          <h5>Criadores</h5>
-          <p>Equipe de Diw G7 - Puc Minas Coração Eucaristico </p>
-        </div>
-      </div>
-    </div>
-  </footer>
-
-<script src="https://code.jquery.com/jquery-3.5.1.slim.min.js"></script>
-<script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.5.3/dist/umd/popper.min.js"></script>
-<script src="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/js/bootstrap.min.js"></script>
-<script src="assets/js/main.js"></script>
-</body>
-</html>
