--- pmg-es-2024-1-ti1-2401100-g7-saude-e-bem-estar/codigo/Prototipo/assets/css/main.css
@@ -1,156 +1,0 @@
-/*Inicio formatação body*/
-body { 
-  margin: 0;
-  padding:0;
-  border:0;
-  font-family: Arial, Helvetica, sans-serif;
-}
-/*Fim formatação body*/
-
-/*Inicio formatação do header*/
-.header {
-  overflow: hidden;
-  background-color: #3685F5;
-  padding: 20px 10px;
-  color: white;
-}
-
-.header a{
-  float: left;
-  text-align: center;
-  padding: 12px;
-  text-decoration: none;
-  font-size: 18px; 
-  line-height: 25px;
-  border-radius: 4px;
-  color: white;
-}
-
-.header a.logo {
-  font-size: 25px;
-  font-weight: bold;
-}
-
-.header a:hover {
-  background-color: #3077DA;
-  color: white;
-}
-
-.header-right {
-  float: right;
-}
-
-@media screen and (max-width: 500px) {
-  .header a {
-    float: none;
-    display: block;
-    text-align: left;
-  }
-
-  .header-right {
-    float: none;
-  }
-}
-/* Fim formatação do header*/
-
-/*Inicio botão inscrição*/
-.btn-verde{
-  background-color: #3649F5;
-  text-align: center;
-  padding: 10px;
-  text-decoration: none;
-  font-size: 18px; 
-  border-radius: 6px;
-  border-width: 0px;
-  color: white;
-}
-
-.btn-verde:hover {
-  background-color: #2F40D8;
-}
-/*Fim botão inscrição*/
-
-.img-saude{
-  width: 35px;
-}
-
-.cards{
-  background-color: #D1FFD4;
-}
-
-/* Inicio carrosel*/
-.carousel {
-  width: 100%;
-  overflow: hidden;
-}
-
-.carousel-inner {
-  display: flex;
-  transition: transform 0.5s ease;
-}
-
-.carousel-item {
-  min-width: 100%;
-}
-
-.carousel-item img {
-  width: 100%;
-  height: auto;
-}
-
-.carousel-btn {
-  cursor: pointer;
-  position: absolute;
-  top: 50%;
-  transform: translateY(-50%);
-  background: rgba(255, 255, 255, 0.5);
-  padding: 10px;
-  border-radius: 50%;
-  transition: background 0.3s ease;
-}
-
-.carousel-btn:hover {
-  background: rgba(255, 255, 255, 0.8);
-}
-
-.prev-btn {
-  left: 10px;
-}
-
-.next-btn {
-  right: 10px;
-}
-/*Fim Carrosel*/
-
-.link{
-  color: #3077DA;
-}
-
-.link:hover{
-  color: #3685F5;
-}
-
-/*Inicio footer*/
-.footer {
-  font-family: Arial, Helvetica, sans-serif;
-  width: 100%;
-  background-color: #3685F5;
-  color: white;
-  bottom: 0;
-  margin-top: 20px;
-}
-/*Fim footer*/
-
-
-.cardTelaEscolha {
-  width: 200px;
-  height: 150px;
-  background-color: #f0f0f0;
-  border: 1px solid #ccc;
-  border-radius: 5px;
-  margin: 10px;
-  padding: 20px;
-  display: inline-block;
-  cursor: pointer;
-  text-align: center;
-}
