--- pmg-es-2024-1-ti1-2401100-g7-saude-e-bem-estar/codigo/Prototipo/assets/css/perfil_med.css
@@ -1,13 +1,0 @@
-.fundo{
-    background-color: #36C8F5;
-    color: white;
-    width: 100%;
-    
-}
-.img_perfil{
-    margin-right:50px ;
-    width: 20%;
-}
-.btn-right {
-    float: right;
-  }
