--- pmg-es-2024-1-ti1-2401100-g7-saude-e-bem-estar/codigo/Prototipo/index.html
@@ -1,163 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link href="assets/css/main.css" rel="stylesheet">
-    <link href="assets/css/bootstrap.min.css" rel="stylesheet">
-    
-    <title>Santé</title>
-</head>
-<body>
-  <!--Cabeçalho do site -->
-    <Header>
-        <div class="header">
-          <a href="index.html" class="logo"> Santé <img src="assets/img/Saude2.png" alt="saude" class="img-saude"></a>
-          <div class="header-right">
-            <a href="#home">Meu Caso</a>
-            <a href="#contact">Quem Somos</a>
-              <button type="button" onclick="window.location.href='Cadastro_Med.html'" class="btn-verde">Cadastrar</button></a>
-              <button href="#" type="button" onclick="window.location.href='Login_Med.html'"class="btn-verde">Login</button></a>
-          </div>
-        </div>
-    </Header>
-
-<!--Conteudo central do site -->
-    <main>
-      <h1 class="text-center">Conecte-se à sua saúde, conecte-se ao futuro, conecte-se conosco.</h1><br>
-
-<!--Carrosel da pagina incial -->
-<div id="carouselExampleIndicators" class="carousel slide " data-ride="carousel">
-  <ol class="carousel-indicators">
-    <li data-target="#carouselExampleIndicators" data-slide-to="0" class="active"></li>
-    <li data-target="#carouselExampleIndicators" data-slide-to="1"></li>
-    <li data-target="#carouselExampleIndicators" data-slide-to="2"></li>
-  </ol>
-  <div class="carousel-inner">
-    <div class="carousel-item active">
-      <img src="assets/img/Carrosel1.jpeg" class="d-block w-25 img-fluid" alt="img1">
-    </div>
-    <div class="carousel-item">
-      <img src="assets/img/Saude.png" class="d-block w-25 img-fluid" alt="img2">
-    </div>
-    <div class="carousel-item">
-      <img src="assets/img/Carrosel1.jpeg" class="d-block w-25 img-fluid" alt="img3">
-    </div>
-  </div>
-  <a class="carousel-control-prev" href="#carouselExampleIndicators" role="button" data-slide="prev">
-    <span class="carousel-control-prev-icon" aria-hidden="true"></span>
-    <span class="sr-only">Previous</span>
-  </a>
-  <a class="carousel-control-next" href="#carouselExampleIndicators" role="button" data-slide="next">
-    <span class="carousel-control-next-icon" aria-hidden="true"></span>
-    <span class="sr-only">Next</span>
-  </a>
-</div>
-
-
-<!--Parte das funcionalidades -->
-
-      <h2>Sempre cuidado de você</h2>
-      <p>Trazemos diversar funcinalidedes que podem te auxiliar na sua jornada</p>
-
-      <div class="row">
-        <div class="col-sm-3 mb-3">
-          <div class="card text-center">
-            <div class="card-body cards">
-              <h5 class="card-title">Calendario de consultas</h5>
-              <p class="card-text">With supporting text below as a natural lead-in to additional content.</p>
-              <a href="Calendario.html" class="btn btn btn-success">Go somewhere</a>
-            </div>
-          </div>
-        </div>
-        <div class="col-sm-3 mb-3">
-          <div class="card text-center">
-            <div class="card-body cards">
-              <h5 class="card-title">Marcar atendimento</h5>
-              <p class="card-text">With supporting text below as a natural lead-in to additional content.</p>
-              <a href="#" class="btn btn btn-success">Go somewhere</a>
-            </div>
-          </div>
-        </div>
-        <div class="col-sm-3 mb-3">
-          <div class="card text-center">
-            <div class="card-body cards">
-              <h5 class="card-title">Agendamento de consultas presencial</h5>
-              <p class="card-text">With supporting text below as a natural lead-in to additional content.</p>
-              <a href="#" class="btn btn btn-success">Go somewhere</a>
-            </div>
-          </div>
-        </div>
-        <div class="col-sm-3 mb-3">
-          <div class="card text-center">
-            <div class="card-body cards">
-              <h5 class="card-title">Consulta de historio para medicos</h5>
-              <p class="card-text">With supporting text below as a natural lead-in to additional content.</p>
-              <a href="#" class="btn btn-success">Go somewhere</a>
-            </div>
-          </div>
-        </div>
-      </div><br>
-
-<!--Parte de noticias do site -->
-
-      <h2>Acompanhe as noticias</h2>
-      <p>As noticias mais recentes sobre a area da saude</p>
-
-      <div class="row">
-        <div class="col-sm-4 mb-3">
-          <div class="card text-center">
-            <img src="assets/img/Carrosel1.jpeg" class="card-img-top" alt="Card1">
-            <div class="card-body cards">
-              <a href="#" class="btn btn btn-success w-50">Esportes</a>
-            </div>
-          </div>
-        </div>
-        <div class="col-sm-4 mb-3">
-          <div class="card text-center">
-            <img src="assets/img/Carrosel1.jpeg" class="card-img-top" alt="Card2">
-            <div class="card-body cards">
-              <a href="#" class="btn btn btn-success w-50">Exames</a>
-            </div>
-          </div>
-        </div>
-        <div class="col-sm-4 mb-3">
-          <div class="card text-center">
-            <img src="assets/img/Carrosel1.jpeg" class="card-img-top" alt="Card3">
-            <div class="card-body cards">
-              <a href="#" class="btn btn btn-success w-50">Pesquisa</a>
-            </div>
-          </div>
-        </div>
-      </div><br>
-    </main>
-
-<!--Footer das informações do site -->
-    <footer>
-      <div class="footer">
-        <div class="row g-0 text-center">
-
-          <div class="col-6 col-md-4 p-2">
-            <h5>Seja Nosso cliente</h5>
-            <p>Cadastre-se</p>
-          </div>
-          <div class="col-6 col-md-4 p-2">
-            <h5>Fale com a gente</h5>
-            <p>Enviar Mensagem</p>
-            <p>Suporte (31) 9 9999-9999</p>
-            <p>Reclame aqui</p>
-          </div>
-          <div class="col-6 col-md-4 p-2">
-            <h5>Criadores</h5>
-            <p>Equipe de Diw G7 - Puc Minas Coração Eucaristico </p>
-          </div>
-        </div>
-      </div>
-    </footer>
-
-  <script src="https://code.jquery.com/jquery-3.5.1.slim.min.js"></script>
-  <script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.5.3/dist/umd/popper.min.js"></script>
-  <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/js/bootstrap.min.js"></script>
-  <script src="assets/js/main.js"></script>
-</body>
-</html>
