--- pmg-es-2024-1-ti1-2401100-g7-saude-e-bem-estar/codigo/README.md
@@ -1,5 +1,0 @@
-# Código do Projeto
-
-Mantenha neste diretório todo o código fonte do projeto. 
-
-Se necessário, descreva neste arquivo aspectos relevantes da estrutura de diretórios criada para organização do código.
