--- pmg-es-2024-1-ti1-2401100-g7-saude-e-bem-estar/docs/apresentacao/README.md
@@ -1,4 +1,0 @@
-# Divulgação: Apresentação do Projeto
-
-Nesta pasta inclua arquivos de slides que foram produzidos para apresentações do projeto e de seus resultados.
-
