--- pmg-es-2024-1-ti1-2401100-g7-saude-e-bem-estar/docs/video/README.md
@@ -1,8 +1,0 @@
-# Vídeos do Projeto
-A relação abaixo lista os vídeos feitos para o projeto:
- - [Vídeo AAAAAAAA]()
- - [Vídeo BBBBBBBB]()
-
-> Nesta pasta inclua arquivos de vídeo produzidos para divulgação do 
-> projeto e seus resutados.
-
